+++ examples/unblocking.rs
@@ -2,16 +2,16 @@ use std::{mem::forget, thread::sleep, time::Duration};
 
 use discord_presence::Client;
 
+mod helpers;
+
 fn main() {
-    tracing_subscriber::fmt()
-        .with_max_level(tracing::Level::TRACE)
-        .init();
+    helpers::logging::init_logging();
 
     let mut client = Client::new(1003450375732482138);
 
     client.start();
 
-    tracing::error!("Due to the way unblocking activity setting works, this example does not seem to work currently (at least on Windows).");
+    log::error!("Due to the way unblocking activity setting works, this example does not seem to work currently (at least on Windows).");
     {
         let ready = client.on_ready({
             let client = client.clone();
@@ -40,7 +40,7 @@ fn main() {
         eprintln!("An error occured, {:?}", ctx.event);
     });
 
-    tracing::trace!("Made it to the final line");
+    log::trace!("Made it to the final line");
 
     // keep the main thread alive
     loop {